+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All Notable changes to `bakame/aide-profiler` will be documented in this file.
 ### Added
 
 - `Label::random` accepts an optional `$length` parameter to adjust the label length.
+- `Marker::complete` and `Marker::isComplete`
+- `Profiler::hasSummaries`
 
 ### Fixed
 
@@ -17,6 +19,7 @@ All Notable changes to `bakame/aide-profiler` will be documented in this file.
 - **BC BREAK:** `Exporter::exportProfilingData` is renamed `Exporter::exportSummary`
 - **BC BREAK:** `Snapshot::toArray` and `Snapshot::jsonSerialize` representation simplified
 - **BC BREAK:** `Summary::toArray` and `Summary::jsonSerialize` representation simplified
+- **BC BREAK:** `Marker::finish` throws an `UnableToProfile` exception previously was an `InvalidArgument` exception when it cannot complete its task.
 - `Snapshot::cpu` keys presence is validated on instantiation
 
 ### Deprecated